viz improvements, added region info dump
The output structure for visualizations is now deterministic.

Fixed a bug where the viz crashed if the input alignment model names
aren't in the name#class/subclass format.

Added a zoomed in view of the reference annotation track.

Added a new --regions <path> flag that dumps region information to a
file. The format is
<region_#>,<target/sequence>,<target_start>:<target_end>,<line_start>:<line_end>